Add --activitysubscribers option
This allows marking of bugs that were last touched by "us", where "us"
is defined by a Launchpad group.

To prevent potential web UI races, we consider any activity not by "us"
up to one hour before "we" touched it as bugs that should not be marked.

This works, but is incredibly slow due to the round-tripping to
Launchpad required on a per-bug basis. In my testing, a typical day's
query goes from 8 seconds to 281 seconds. So I don't think this option
should be default. We'll see how useful it is over time I think.

Note that the speed issue should not degenerate the further we get into
the past: it should be proportional to the number of bugs in the result
when used without --activitysubscribers only.