Add testing via tox and packaging
This commit renames the main executable to 'ustriage' in order to
have a shorter command to use. It also adds in a setup.py to
allow for the packaging up of the script as well as for use by tox
to enable testing via flake8 and pylint.

Additional changes due to test results were also made.