Blacklist some packages by default
Hardcode some packages that we don't need to triage as part of the usual
triage rota because they get looked after by separate efforts.

Add a --no-blacklist option to disable this behaviour.

This also moves the printing of the number of bugs found to the
print_bugs() function, which is necessary to keep the count accurate
against the filter. Otherwise the counting would happen before the
filter is applied.

I think print_bugs is an OK place to place the filter for now. It would
be better if we could do it closer to the source, but I'm not aware of
a Launchpad API mechanism to do it closer to the search query, and
filtering afterwards needs the Task objects created to be able to infer
the source package name from the tasks correctly.