Include results for all relevant bug statuses
The Launchpad API is unclear on which bug statuses are filtered out by
default, so some were getting missed.

For example, if a bug was incorrectly marked Fix Released and a reporter
commented on how it wasn't without changing the status, our triage
process wouldn't flag it.

We can be more explicit about this, and make sure that everything we
want is included.