Fix: Speed up activitysubscribers handling
After waiting too long too often I've checked where time is lost and
it came down to 0.1-0.5 seconds per round trip on comment analysis
for activitysubscribers. We had bugs with 40 or 51 comments which
thereby took around 20 seconds each.

But I realized that the collection we get is already ordered.
Therefore we can just slice and only process the few latest updates
for the very same results at a faster execution time.
Example 23 entries and the last three slices are
 https://api.launchpad.net/devel/ubuntu/+source/mysql-8.0/+bug/1882527/comments/20
 https://api.launchpad.net/devel/ubuntu/+source/mysql-8.0/+bug/1882527/comments/21
 https://api.launchpad.net/devel/ubuntu/+source/mysql-8.0/+bug/1882527/comments/22
No need to fetch ALL to (re-)order them by date.

This speeds up execution with long lists and busy bugs a lot with
an average of ~1.7 seconds per bug now. In our usual length of 10-30
bugs in sevrer-next or bug triage that means 40 seconds instead of the
sometimes rather long multi-minute waits that we had.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>