UI: Order bug-scrub Backlog by date
Usually we order by last-team-touched,bugnumber,source-package-name.
But in the newly used backlog views it is much more useful to have it
ordered by last-updated.

This will have to probe the last updated date once per task in the list
to order and only then can we reduce the list e.g. via limit_backlog.
Therfore this can take a while. Currently 485 bugs * 0.3 seconds a bit
more than two minutes.

I have experimented with order_by on searchTasks to avoid the local
reordering, but so far it wasn't reliable enough to use it.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>