UI: convert --bug-scrub into per-output switches
The default is and will stay the triage content. But we can now
disable that (--show-no-triage) and for bug scrub show these
(--show-backlog --show-tagged) or use any combination thereof.

The new modes are generally also useful to show any tagged and any
subscribed list of bugs via --lpname and --tag-next and I've updated
the cli argument help texts to show that.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>