Use tags as list internally more consistently
Fixes BadRequest exception when calling print_tagged_bugs() with no
defined expiration.

The Launchpad API accepts passing tags as either a list or a single
string.  Prior to 0a8e5e, the ustriage code assumed a single tag
internally, and passed it to Launchpad directly, however
print_tagged_bugs() was an exception in that it passes the tag as a
list (since it also adds '-bot-stop-nagging').

With commit 0a8e5e and especially b2624f, create_bug_list() is modified
to expect a list of tags rather than a single tag.  Unfortunately these
changes overlooked that print_tagged_bugs() was already passing tag as a
list of tags, which resulted in passing Launchpad a tags argument
comprised of a list inside a list.

This commit corrects the whole situation by treating tags as a list
uniformly throughout all of the internal code.  However, note that the
user interface remains unchanged: '--tag foo' still accepts only a
single tag as a string.