Defaults for age reporting
Our common use cases are triage and tag/list checks. Those come
with different demands for default age checking.

Triage is all-recent so it comes disabled in this case.
Tag/List checks OTOH are usually done in a weekly pattern, so we check
if something happened in the last week.

Either of that can be overwritten by the new --age argument.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>