refactor: separate opening of bugs in the browser
This simplification refactoring dictaed by tox wasn't bad after all.

We had some nice but not important changes:
- back in the recommended limit of branches/statements
- reducing variables (if the browser was already opened is tracked
  in the control variable, which now also allows to check how many we
  opened - this isn't used yet but would allow e.g. a limit later)

More important was that along the way it was found that --open did not
work for --show-tagged and --show-subscribed which used --open-exp
instead (wrong). This is now fixed and only the listing of
tagged/subscribed under the expired banner will use --open-exp while
the other tagged/subscribed listings use --open now (as one would
expect).

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>