task: report brief per release status
In the housekeeping sessions we often wished for seeing which
releases are marked as affected and in which state they are.

This change will report that just after the flag section,
if a character is present that release is marked as affected.
Lower case = open, upper case = closed.

Examples:
Fixed in devel, open in jammy => Dj
Fixed in devel, open in bionic, fixed in Focal+Impish => DbFI

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>