ustriage: detect uploaded SRUs
Thanks to Robie we are able to detect if a bug in a series for a package
has an upload in the SRU queue. Integrated that code to reflect it in
the per-release flags.

By that we exceed what we can express meaningfully with lower/upper case
so it switched from lower/upper to colored characters.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>