readability: avoid conflict of control characters
Strikethrough (postponed) and Colors (release/sru status) did
conflict and in the postponed bugs one could see the literal control
characters. Since postponed is just meant as a hint it is fine to
go with non-format other than the strikethough.

To achieve that we filter the text before applying strikethrough.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>