task: Use the bug task url to instantiate object
Cache some properties that can be parsed from the bug task's url.  This
derives from logic that is used when evaluating status of sibling tasks
to make use of the info in the object itself.

Note that theoretically this could differ from the information if
queried from Launchpad.  The latter would be more canonical, however
doing it this way should be more performant, at least for ustriage's use
case.